requirements.txt
addressable==1.4.2
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
entrypoints==0.2.3
facebookinsights==0.3.4
facepy==1.0.9
Flask==1.0.2
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
keyring==13.2.0
MarkupSafe==1.0
numpy==1.14.5
oauthlib==2.1.0
pandas==0.23.1
pyasn1==0.4.3
pyasn1-modules==0.2.2
pylev==1.3.0
python-dateutil==2.7.3
pytz==2018.5
rauth==0.7.3
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
selenium==3.13.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
XlsxWriter==1.0.5
xlwt==1.3.0
